refactor(minishell.h, lexer.c, lexer_utils2.c): remove ft_lstdupp function and simplify create_node_with_redir function
The ft_lstdupp function was removed because it was unnecessarily duplicating tokens. Instead, the create_node_with_redir function now directly assigns the lexer's redirect to the token's redir and sets lexer's redirect to NULL. This change simplifies the code and reduces memory usage.

Also, the visibility of several functions in lexer.c was changed to static to limit their scope to the file, improving encapsulation and preventing potential naming conflicts.