fix: fixed error with command not found and pipe
Due to the check of the validity of the command before the fork, this was
error prone in precise case so to avoid this, we always fork and check
in the child process if the command is valid to send to execve. Due to
this modification we don't need anymore test in the loop when waiting
all process so I take the older version of the launch_execution (don't
need to split anymore due to more than 25 lines).